+++ airflow-core/src/airflow/ui/src/layouts/Nav/Nav.tsx
@@ -17,7 +17,6 @@
  * under the License.
  */
 import { Box, Flex, VStack, useDisclosure } from "@chakra-ui/react";
-import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { FiDatabase, FiHome, FiClock } from "react-icons/fi";
 import { Link } from "react-router-dom";
@@ -28,11 +27,9 @@ import {
   usePluginServiceGetPlugins,
 } from "openapi/queries";
 import type { ExternalViewResponse } from "openapi/requests/types.gen";
-import { AirflowPin } from "src/assets/AirflowPin";
 import { DagIcon } from "src/assets/DagIcon";
-import { useColorMode } from "src/context/colorMode";
+import { Logo } from "src/components/Logo";
 import { useTimezone } from "src/context/timezone";
-import { useConfig } from "src/queries/useConfig";
 import { getTimezoneOffsetString, getTimezoneTooltipLabel } from "src/utils/datetimeUtils";
 import type { NavItemResponse } from "src/utils/types";
 
@@ -103,15 +100,6 @@ export const Nav = () => {
   const { selectedTimezone } = useTimezone();
   const offset = getTimezoneOffsetString(selectedTimezone);
   const tooltipLabel = getTimezoneTooltipLabel(selectedTimezone);
-  const theme = useConfig("theme") as unknown as { icon?: string; icon_dark_mode?: string } | undefined;
-
-  // Custom icons handling
-  const { colorMode } = useColorMode();
-  const darkIcon = theme?.icon_dark_mode ?? undefined;
-  const lightIcon = theme?.icon ?? undefined;
-  const iconSrc = colorMode === "dark" && darkIcon !== undefined ? darkIcon : lightIcon;
-  const hasIconSrc = Boolean(iconSrc);
-  const [failedLoadingCustomIcon, setFailedLoadingCustomIcon] = useState({ dark: false, light: false });
 
   // Get both external views and react apps with nav destination
   const navItems: Array<NavItemResponse> =
@@ -169,25 +157,15 @@ export const Nav = () => {
       <Flex alignItems="center" flexDir="column" gap={1} width="100%">
         <Box alignItems="center" asChild boxSize={14} display="flex" justifyContent="center">
           <Link title={translate("nav.home")} to="/">
-            {hasIconSrc && colorMode && !failedLoadingCustomIcon[colorMode] ? (
-              // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions
-              <img
-                alt="Airflow"
-                onError={() => setFailedLoadingCustomIcon((prev) => ({ ...prev, [colorMode]: true }))}
-                src={iconSrc}
-                style={{ height: "2rem", width: "2rem" }}
-              />
-            ) : (
-              <AirflowPin
-                _motionSafe={{
-                  _hover: {
-                    transform: "rotate(360deg)",
-                    transition: "transform 0.8s ease-in-out",
-                  },
-                }}
-                boxSize={8}
-              />
-            )}
+            <Logo
+              _motionSafe={{
+                _hover: {
+                  transform: "rotate(360deg)",
+                  transition: "transform 0.8s ease-in-out",
+                },
+              }}
+              boxSize={8}
+            />
           </Link>
         </Box>
         <NavButton data-testid="nav-home-link" icon={FiHome} title={translate("nav.home")} to="/" />
